# perf: fast regex check on string