--- a/lib/serializer.js
+++ b/lib/serializer.js
@@ -1,7 +1,7 @@
 'use strict'
 
 // eslint-disable-next-line
-const STR_ESCAPE = /[\u0000-\u001f\u0022\u005c\ud800-\udfff]|[\ud800-\udbff](?![\udc00-\udfff])|(?:[^\ud800-\udbff]|^)[\udc00-\udfff]/
+const STR_ESCAPE = /[\u0000-\u001f\u0022\u005c\ud800-\udfff]/
 
 module.exports = class Serializer {
   constructor (options) {