@cesco69 2024-03-15
Side node, also merged into the suorce project of the regex https://github.com/BridgeAR/fast-json-escape/pull/6