@cesco69 2024-03-18
> What attention do you need?

@mcollina Because it seem a good improvement on long string parsing (from 2.2% to 5.8), I would like to get this PR into the master's as quickly as possible.